third_party/WebKit/LayoutTests/fast/canvas/drawImage-with-broken-image.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<script>
    window.jsTestIsAsync = true;
    description("This test checks behavior of Canvas::drawImage with a broken source image.");

    var InvalidStateError = "InvalidStateError: Failed to execute 'drawImage' on 'CanvasRenderingContext2D': The source width is 0.";
    var TypeMismatchError = "TypeMismatchError: The type of an object was incompatible with the expected type of the parameter associated to the object.";

    // Create an image with invalid data.
    var invalidImage = new Image();
    invalidImage.src = 'resources/repaint.js';
    invalidImage.onerror = draw;

    var ctx = document.createElement("canvas").getContext('2d');
    function draw() {
        // null images should throw TypeError
        shouldThrow("ctx.drawImage(null, 0, 0)");
        shouldThrow("ctx.drawImage(null, 0, 0, 20, 20)");
        shouldThrow("ctx.drawImage(null, 0, 0, 20, 20, 0, 0, 20, 20)");

        // broken images should not throw
        shouldBe("ctx.drawImage(invalidImage, 0, 0)", "undefined");
        shouldBe("ctx.drawImage(invalidImage, 0, 0, 20, 20)", "undefined");
        shouldBe("ctx.drawImage(invalidImage, 0, 0, 20, 20, 0, 0, 20, 20)", "undefined");
        shouldBe("ctx.drawImage(invalidImage, 0, 0, 0, 20)", "undefined");
        shouldBe("ctx.drawImage(invalidImage, 0, 0, 0, 20, 0, 0, 20, 20)", "undefined");

        finishJSTest();
    }
</script>
</body>
</html>
